Part 3
1) When the game starts, the player should see a blank scene.
2) Every createBronzeTime seconds, a bronze ore (i.e. a red cube)
appears.
a. Note: It should be obvious to the player that more ore is appearing
over time, so don't put the new bronze ore in the same place as existing
ore, and ensure the new ore appears someplace the player can see it.
3) After switchToSilverTime seconds, instead of spawning bronze ore,
spawn silver ore (i.e. a white cube).
4) After an additional stopSpawningTime seconds, stop spawning ore.