part four
1) When the player clicks on an ore, it should disappear.
2) Instead of what's described previously, use the following algorithm
to spawn the ore:
a. Spawn an ore every spawnTime seconds.
b. If there are <4 bronze ore, spawn a bronze.
c. If there are 4+ bronze, spawn a silver.
d. If there are exactly 2 bronze and 2 silver, spawn a gold (yellow
cube) instead (but just spawn 1 gold and then go back to the normal
rules.)
3) Whenever the player clicks an ore, give the player points as follows:
a. bronzePoints for a bronze
b. silverPoints for a silver
c. goldPoints for a gold